Move Linux specific code into /src/linux
