+++ Cargo.toml
@@ -10,9 +10,11 @@ readme = "README.md"
 
 [dependencies]
 derive_builder = "0.7.1"
+thiserror = "1.0"
+
+[target.'cfg(target_os = "linux")'.dependencies]
 libc = "0.2.66"
 neli = "0.4.3"
-thiserror = "1.0"
 
 [dev-dependencies]
 anyhow = "1.0"
