+++ src/linux/err/connect_error.rs
