+++ src/linux/err/link_device_error.rs
