+++ src/linux/err/set_device_error.rs
