+++ src/linux/socket/list_device_names_utils.rs
@@ -1,5 +1,5 @@
 use super::parse::parse_nla_nul_string;
-use crate::err::ListDevicesError;
+use crate::linux::err::ListDevicesError;
 use neli::consts::{Arphrd, Ifla, NlmF, Nlmsg, Rtm};
 use neli::nl::Nlmsghdr;
 use neli::rtnl::Ifinfomsg;
@@ -47,7 +47,7 @@ pub fn parse_ifinfomsg(
 
                 if linkinfo.rta_type == libc::IFLA_INFO_KIND {
                     let info_kind = parse_nla_nul_string(&linkinfo.rta_payload)?;
-                    if info_kind == crate::consts::WG_GENL_NAME {
+                    if info_kind == crate::linux::consts::WG_GENL_NAME {
                         is_wireguard = true;
                     }
                 }
