+++ src/linux/socket/wg_socket.rs
@@ -1,13 +1,13 @@
-use crate::attr::WgDeviceAttribute;
-use crate::cmd::WgCmd;
-use crate::consts::{WG_GENL_NAME, WG_GENL_VERSION};
-use crate::err::{ConnectError, GetDeviceError, SetDeviceError};
-use crate::get;
-use crate::set;
-use crate::set::create_set_device_messages;
-use crate::socket::parse::*;
-use crate::socket::NlWgMsgType;
-use crate::DeviceInterface;
+use crate::linux::attr::WgDeviceAttribute;
+use crate::linux::cmd::WgCmd;
+use crate::linux::consts::{WG_GENL_NAME, WG_GENL_VERSION};
+use crate::linux::err::{ConnectError, GetDeviceError, SetDeviceError};
+use crate::linux::get;
+use crate::linux::set;
+use crate::linux::set::create_set_device_messages;
+use crate::linux::socket::parse::*;
+use crate::linux::socket::NlWgMsgType;
+use crate::linux::DeviceInterface;
 use libc::IFNAMSIZ;
 use neli::consts::{NlFamily, NlmF, Nlmsg};
 use neli::genl::Genlmsghdr;
